pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "livekit-flows"
version = "0.1.0"
description = "A library for building declarative agent flows with LiveKit."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.2.8",
    "pydantic>=2.11.7",
    "aiohttp>=3.12.15",
    "jinja2>=3.1.6",
]

[project.urls]
Homepage = "https://github.com/mateuszkulpa/livekit-flows"
Repository = "https://github.com/mateuszkulpa/livekit-flows"

[tool.hatch.build.targets.wheel]
packages = ["livekit_flows"]

[tool.uv]
default-groups = ["dev", "examples"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-aiohttp>=1.1.0",
    "ruff>=0.12.11",
    "ty>=0.0.1a19",
]
examples = [
    "livekit-agents>=1.2.8",
    "livekit-plugins-openai>=1.2.8",
    "livekit-plugins-cartesia>=1.2.8",
    "livekit-plugins-deepgram>=1.2.8",
    "livekit-plugins-silero>=1.2.8",
    "python-dotenv>=1.1.1",
]
